requirements.txt
requests>=2.31.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pycryptodome>=3.19.0
base64url>=1.0.0
rich>=13.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
